Trying to fix client loop when server & client app at different machines

+++ srv/src/UserBased.cpp
@@ -74,7 +74,7 @@ void UserBase::addUsers(User& newUser)
 
 int UserBase::getUserCount()
 {
-	return usrBase->size();
+	return (int)usrBase->size();
 }
 
 void UserBase::saveUsrBase()
